FIx
Fix pix payment.
